app-history/navigate/navigate-unload.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  let iframe_exception = i.contentWindow.DOMException;
  window.onload = t.step_func(() => {
    i.contentWindow.onunload = t.step_func(() => {
      promise_rejects_dom(t, 'AbortError', iframe_exception, i.contentWindow.appHistory.navigate("#")).then(() => t.done());
    });
    i.contentWindow.appHistory.navigate("?1")
  });
}, "Navigate inside onunload");
</script>
